refactor auth services
