+++ src/app/api/auth/users/route.ts
@@ -2,7 +2,7 @@
 import { NextResponse } from "next/server";
 
 import type { UserInput } from "@/types/user";
-import { createUser } from "@/lib/services/auth/createUser";
+import { createUser } from "@/lib/services/auth/users/createUser";
 
 export async function POST(request: Request) {
   try {
